+++ commands/assign_stemcell.go
@@ -2,6 +2,7 @@ package commands
 
 import (
 	"fmt"
+	"os"
 	"strings"
 
 	"github.com/pivotal-cf/jhanda"
@@ -12,7 +13,7 @@ type AssignStemcell struct {
 	logger  logger
 	service assignStemcellService
 	Options struct {
-		ConfigFile      string `long:"config"   short:"c"  description:"path to yml file for configuration (keys must match the following command line flags)"`
+		interpolateConfigFileOptions
 		ProductName     string `long:"product"  short:"p"  description:"name of Ops Manager tile to associate a stemcell to" required:"true"`
 		StemcellVersion string `long:"stemcell" short:"s"  description:"associate a particular stemcell version to a tile." default:"latest"`
 	}
@@ -41,7 +42,7 @@ func (as AssignStemcell) Usage() jhanda.Usage {
 }
 
 func (as AssignStemcell) Execute(args []string) error {
-	err := loadConfigFile(args, &as.Options, nil)
+	err := loadConfigFile(args, &as.Options, os.Environ)
 	if err != nil {
 		return fmt.Errorf("could not parse assign-stemcell flags: %s", err)
 	}