+++ commands/config_template.go
@@ -14,10 +14,7 @@ type ConfigTemplate struct {
 	environFunc   envProvider
 	buildProvider buildProvider
 	Options       struct {
-		ConfigFile string   `long:"config"                     short:"c" description:"path to yml file for configuration (keys must match the following command line flags)"`
-		VarsEnv    []string `long:"vars-env" env:"OM_VARS_ENV"           description:"load variables from environment variables matching the provided prefix (e.g.: 'MY' to load MY_var=value)"`
-		VarsFile   []string `long:"vars-file"                  short:"l" description:"load variables from a YAML file"`
-		Vars       []string `long:"var"                        short:"v" description:"Load variable from the command line. Format: VAR=VAL"`
+		interpolateConfigFileOptions
 
 		PivnetApiToken    string `long:"pivnet-api-token"`
 		PivnetProductSlug string `long:"pivnet-product-slug"                          description:"the product name in pivnet"`