+++ commands/configure_authentication.go
@@ -20,7 +20,8 @@ type ConfigureAuthentication struct {
 	logger      logger
 	environFunc func() []string
 	Options     struct {
-		ConfigFile             string   `long:"config"                short:"c"                                 description:"path to yml file for configuration (keys must match the following command line flags)"`
+		interpolateConfigFileOptions
+
 		Username               string   `long:"username"              short:"u"  env:"OM_USERNAME"              description:"admin username" required:"true"`
 		Password               string   `long:"password"              short:"p"  env:"OM_PASSWORD"              description:"admin password" required:"true"`
 		DecryptionPassphrase   string   `long:"decryption-passphrase" short:"dp" env:"OM_DECRYPTION_PASSPHRASE" description:"passphrase used to encrypt the installation" required:"true"`
@@ -28,9 +29,6 @@ type ConfigureAuthentication struct {
 		HTTPSProxyURL          string   `long:"https-proxy-url"                                                 description:"proxy for outbound HTTPS network traffic"`
 		NoProxy                string   `long:"no-proxy"                                                        description:"comma-separated list of hosts that do not go through the proxy"`
 		PrecreatedClientSecret string   `long:"precreated-client-secret"                                        description:"create a UAA client on the Ops Manager vm. The client_secret will be the value provided to this option"`
-		VarsEnv                []string `long:"vars-env" env:"OM_VARS_ENV"                                      description:"load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)"`
-		VarsFile               []string `long:"vars-file"             short:"l"                                 description:"load variables from a YAML file"`
-		Vars                   []string `long:"var"                   short:"v"                                 description:"load variable from the command line. Format: VAR=VAL"`
 	}
 }